Micro-optimise checked arithmetic in huffman.
Motivation:

Huffman decoding has rapidly become one of the most expensive parts of
the program. Some of the code here is unavoidably expensive, but there
is some unnecessary checked arithmetic here on offset. Removing that
provides a small but meaningful improvement to the performance of this
function.

Modifications:

- Remove the checked arithmetic on offset, which is bounded by a
   previous operation that was already checked.

Result:

This function gets cheaper by about 4%, though the benchmark as a whole
only sees a 0.1% improvement.