Content-Length errors shouldn't crash
Motivation:

When we added new content length verification and new errors, we didn't
update the places that were handling the old error. This code updates
that logic to ensure that we don't crash.

Modifications:

Add new internal protocol to which the content length validation errors
conform.
Catch that protocol instead.

Result:

No crashes on invalid content-length