HTTP/2 NIOAsyncChannel stock pipeline config
Motivation:

This continues the work to expose functionality which allows users to interact
with HTTP/2 connections via async abstractions and using structured
concurrency.

We build atop of previous work to configure pipelines to deal with
HTTP/2 (with no protocol negotiation) and to wrap connection/stream
channels with `NIOAsyncChannel`s. This will allow users to iterate over
streams and HTTP2Frames.

This PR exposes conveniences for configuring HTTP/2 pipelines in the
simplest case where the user knows that any inserted channel handlers
will not change the `InboundOut`/`OutboundIn` types of the pipeline
handlers.

Modifications:

Provide functions which specialize configuration functions which exposed
generic `NIOAsyncChannel`s with standard `HTTP2Frame` types..

Result:

Users will be able to create and interact with HTTP/2 connections via
`NIOAsyncChannel`s by using simpler configuration functions.
Because HTTP/2 is a negotiated protocol and we do not yet handle it,
this is of limited utility.