HTTP/2 pipeline NIOAsyncChannel pipeline config (#403)
* HTTP/2 pipeline NIOAsyncChannel pipeline config

Motivation:

This continues the work to expose functionality which allows users to interact
with HTTP/2 connections via async abstractions and using structured
concurrency.

We build atop of previous work to configure pipelines to deal with
HTTP/2 (with no protocol negotiation) and to wrap connection/stream
channels with `NIOAsyncChannel`s. This will allow users to iterate over
streams and HTTP2Frames.

Modifications:

Provide functions which configure channels and pipelines with the HTTP2
handler, expose a multiplexer for dealing with streams and wrap
connection and stream channels with `NIOAsyncChannel`s.

Result:

Users will be able to create and interact with HTTP/2 connections via
`NIOAsyncChannel`s. Because HTTP/2 is a negotiated protocol and we do
not yet handle it, this is of limited utility.

* review comments

* Update Sources/NIOHTTP2/HTTP2ChannelHandler+InlineStreamMultiplexer.swift

Co-authored-by: George Barnett <gbarnett@apple.com>

---------

Co-authored-by: George Barnett <gbarnett@apple.com>