regularize createStreamChannel sendability
Calls to `createStreamChannel` now always go via a `SendableView` lens
type which serializes access on an event loop with an unconditional hop.
This combines the serialization with the unconditional hop which was
previously performed at the `HTTP2CommonInboundStreamMultiplexer` level,
in some cases saving a hop.

The `SendableView` is performed at the higher multiplexer levels rather
than at the 'common' level because the higher levels pass in a `self` to
lower calls which would otherwise need to be transferred in to the event
loop hop.