Accept PING frames even when fully quiesced
Motivation:

When the connection is fully quiesced, an endpoint responds with a connection error when it receives a PING frame. This behaviour is unnecessary.

Modifications:

- Adjust `HTTP2ConnectionStateMachine` to accept PING frames even when the connection is fully quiesced.

Result:

When the connection is fully quiesced, an endpoint will treat the receipt of a PING frame as it would when the connection is active.